testpmd: Move to a new package
Also change the file name to console.go
